add admin folder

+++ routes/web.php
@@ -16,12 +16,12 @@ Route::middleware('auth')->group(function(){
     Route::get('/dashboard', [PageController::class, 'dashboard'])->name('dashboard.index');
 
     // leaveType (new)
-    Route::get('/leaveType/create', [PageController::class, 'leaveType_create'])->name('leaveType.create');
-    Route::post('/leaveType', [LeaveTypeController::class, 'store'])->name('leaveType.store');
-    Route::get('/leaveType/index', [LeaveTypeController::class, 'index'])->name('leaveType.index');
-    Route::get('/leaveType/{id}/edit', [LeaveTypeController::class, 'edit'])->name('leaveType.edit');
-    Route::put('/leaveType/{id}', [LeaveTypeController::class, 'update'])->name('leaveType.update');
-    Route::delete('/leaveType/{id}', [LeaveTypeController::class, 'destroy'])->name('leaveType.destroy');
+    Route::get('/admin/leaveType/create', [PageController::class, 'leaveType_create'])->name('admin.leaveType.create');
+    Route::post('/admin/leaveType', [LeaveTypeController::class, 'store'])->name('admin.leaveType.store');
+    Route::get('/admin/leaveType/index', [LeaveTypeController::class, 'index'])->name('admin.leaveType.index');
+    Route::get('/admin/leaveType/{id}/edit', [LeaveTypeController::class, 'edit'])->name('admin.leaveType.edit');
+    Route::put('/admin/leaveType/{id}', [LeaveTypeController::class, 'update'])->name('admin.leaveType.update');
+    Route::delete('/admin/leaveType/{id}', [LeaveTypeController::class, 'destroy'])->name('admin.leaveType.destroy');
 });
 
 Route::middleware('guest')->group(function(){
